moved AntivirusInfo file out of models folder
